Added the ci-cd pipeline and fixed errors
